fix: changes get_readings range
changes get_Readings range from 2 hours to 1. 

The resulting JSON from Canary lists 20 items, the first 12 are air quality, the next 8 are humidity. Temperature is not returned in the first 20 records. I don't know if Canary changed the order in which the data is returned (date-time vs sensor_type) but it appears to be sensor-type, then date desc.

This appears to be why temperature is missing. Changing the window of time to look at to 1 hour resulted in all three data types (air_quality, humidify and temperature) to be retuned in the Api.get_readings call.